adding confirmation to item deletion

+++ src/app/add/morsel/item/item.js
@@ -10,16 +10,21 @@ angular.module( 'Morsel.add.item', [])
     },
     link: function(scope, element, attrs) {
       scope.deleteItem = function() {
+        var confirmed = confirm('Are you sure you want to delete this item?');
+
         scope.deletingItem = true;
 
-        ApiItems.deleteItem(scope.item.id).then(function() {
+        if(confirmed) {
+          ApiItems.deleteItem(scope.item.id).then(function() {
+            //tell morsel to remove item
+            scope.$emit('add.item.delete', scope.item.id);
+          }, function(resp) {
+            scope.deletingItem = false;
+            HandleErrors.onError(resp.data, scope.form);
+          });
+        } else {
           scope.deletingItem = false;
-          //tell morsel to remove item
-          scope.$emit('add.item.delete', scope.item.id);
-        }, function() {
-          scope.deletingItem = false;
-          HandleErrors.onError(resp.data, scope.form);
-        });
+        }
       };
     },
     templateUrl: 'app/add/morsel/item/item.tpl.html'
